pyproject.toml
[tool.poetry]
name = "todos"
version = "0.1.0"
description = ""
authors = ["saadalamgir18 <saadalamgir18@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
sqlmodel = "^0.0.18"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"

[tool.poetry.scripts]
dev = "todos.main:start"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
